+++ EntraAuth/internal/functions/authentication/Connect-ServiceRefreshToken.ps1
@@ -18,24 +18,68 @@
 	#>
 	[CmdletBinding()]
 	param (
-		[Parameter(Mandatory = $true)]
-		$Token
+		[Parameter(Mandatory = $true, ParameterSetName = 'Token')]
+		$Token,
+
+		[Parameter(Mandatory = $true, ParameterSetName = 'Details')]
+		[string]
+		$RefreshToken,
+
+		[Parameter(Mandatory = $true, ParameterSetName = 'Details')]
+		[string]
+		$TenantID,
+
+		[Parameter(Mandatory = $true, ParameterSetName = 'Details')]
+		[string]
+		$ClientID,
+
+		[Parameter(Mandatory = $true, ParameterSetName = 'Details')]
+		[string]
+		$Resource,
+
+		[Parameter(ParameterSetName = 'Details')]
+		[string[]]
+		$Scopes = '.default',
+
+		[Parameter(Mandatory = $true, ParameterSetName = 'Details')]
+        [string]
+		$AuthenticationUrl
 	)
 	process {
-		if (-not $Token.RefreshToken) {
-			throw "Failed to refresh token: No refresh token found!"
-		}
+		switch ($PSCmdlet.ParameterSetName) {
+			'Token' {
+				if (-not $Token.RefreshToken) {
+					throw "Failed to refresh token: No refresh token found!"
+				}
+		
+				$effectiveScopes = $Token.Scopes
+		
+				$body = @{
+					client_id = $Token.ClientID
+					scope = @($effectiveScopes).ForEach{"$($Token.Audience)/$($_)"} -join " "
+					refresh_token = $Token.RefreshToken
+					grant_type = 'refresh_token'
+				}
+				$uri = "$($Token.AuthenticationUrl)/$($Token.TenantID)/oauth2/v2.0/token"
+				$authResponse = Invoke-RestMethod -Method Post -Uri $uri -Body $body
+				$Token.SetTokenMetadata((Read-AuthResponse -AuthResponse $authResponse))
+			}
+			'Details' {
+				$effectiveScopes = foreach ($scope in $Scopes) {
+					if ($scope -like "$Resource*") { $scope }
+					else { "$Resource/$scope" }
+				}
 
-		$scopes = $Token.Scopes
-
-		$body = @{
-			client_id = $Token.ClientID
-			scope = @($scopes).ForEach{"$($Token.Audience)/$($_)"} -join " "
-			refresh_token = $Token.RefreshToken
-			grant_type = 'refresh_token'
+				$body = @{
+					client_id = $ClientID
+					scope = $effectiveScopes -join " "
+					refresh_token = $RefreshToken
+					grant_type = 'refresh_token'
+				}
+				$uri = "$($AuthenticationUrl)/$($TenantID)/oauth2/v2.0/token"
+				$authResponse = Invoke-RestMethod -Method Post -Uri $uri -Body $body
+				Read-AuthResponse -AuthResponse $authResponse
+			}
 		}
-		$uri = "$($Token.AuthenticationUrl)/$($Token.TenantID)/oauth2/v2.0/token"
-		$authResponse = Invoke-RestMethod -Method Post -Uri $uri -Body $body
-		$Token.SetTokenMetadata((Read-AuthResponse -AuthResponse $authResponse))
 	}
 }